* PERQemu.csproj: * Trace.cs: * ExecCommands.cs: * Conversion.cs: * ExecutionController.cs: Move/tweak/rename things as the reorganization continues...
* Paths.cs: Updates to Paths.cs.
Merge/update the existing one.

* ExecCommands.cs:
* ExecutionController.cs: Bring in the PERQolator Controller.
Lets us move a ton of stuff out of PERQSystem, which will just
encapsulate the virtual PERQ - instantiated from a given Configuration
by the ExecutionController which will manage the changes in run state.
This is the CLI interface;  the WinForms FrontPanel GUI would hook in
here if it hadn't been obsoleted. Sigh.

* Settings.cs: A simple way to save user preferences.
CLI for loading, setting and saving preferences. Will also remember the
last configuration selected and make that the default loaded on startup.

* Log.cs: Alternate spin on Trace.cs.
Will likely go away as I reconcile the two. Or just switch to the C# equivalent
of Log4j or some generic thing. "I don't know where it's gonna go." -- C. Davis

* PERQ_Chart.jpg: What is your quest?
A kewl dot-matrix printout from the mid-'80s of all the "official" PERQ types.
Someday PERQemu ought to be able to support and run all of these. :-)

* ConfigTypes.cs:
* Configurator.cs:
* Configuration.cs:
* ConfigCommands.cs: Pull in the PERQolator configurator.
Will take a bit of cleanup to integrate, but this is a CLI-driven configuration
tool for defining, loading and saving the setup for a virtual PERQ.