Something needs to be committed and it ain't this This giant sprawling mess is a snapshot of a bunch of changes that got completely out of hand.  Don't worry, it's always darkest before it goes completely black.
* Log.cs:
* Trace.cs:
* Events.cs:
* DebugCommands.cs:
* Z80DebugCommands.cs:
* ExecutionController.cs:
* HighResolutionTimer.cs: Checkpoint for changes to Controller.
Turns out beating on things with hammers may not be the best approach.
Without the GUI front end, trying to scale back the event-driven interplay
between the ExecutionController, PERQSystem and the CLI ends up being
far too convoluted. Added, removed, reworked, and pounded oblong pegs
into triangular voids, ending up with this misshapen lump. All of it needs to
be done over. Everything sucks, all the time, forever.

* Controller.txt:
* NewIOPorts.txt:
* Configuration.txt: Document the unfolding tragedy, triumph?

* CPU.cs:
* CPUBoard.cs:
* PERQ1.cs:
* PERQ1A.cs:
* PERQ24.cs:
* PERQ24A.cs:
* RasterOp.cs:
* ControlStore.cs:
* RegisterFile.cs: Tweaks and refactoring.
Updates to the CPU to support asynchronous operation (on a separate thread)
and minor performance tweaks.

* CIO.cs:
* EIO.cs:
* IOB.cs:
* OIO.cs:
* IOBus.cs:
* IOBoard.cs:
* IZ80System.cs:
* OptionBoard.cs: Refactor for runtime-configurable board types.

* Z80DMA.cs:
* Keyboard.cs:
* PERQFifo.cs:
* Z80IOBus.cs:
* Z80System.cs:
* Z80MemoryBus.cs: Refactoring for configurability, rate limiting.
Tweaks and reorganizing so that the Z80system can be integrated with CIO,
EIO/NIO as they are developed. Konamiman's default ClockSynchronizer is
insane (decimal? seriously?) and doesn't seem to work; use the SystemTimer
heartbeat for accurate Z80 rate limiting like the main CPU.

* MemoryBoard.cs: Rename file to match the class name.

* PERQSystem.cs: Ongoing refactoring.
Now instantiates a CPUBoard that contains the processor and scheduler.
Redo the execution loop to accommodate both asynchronous and
synchronous modes for the CPU and the Z80 threads. Bash on the
ExecutionController interface. It's gonna get worse, before it gets better...

* Scheduler.cs: Embrace the ugly.
Instead of distributing the ugly in the execution loop, concentrate
it here instead! Eh, doesn't seem all that ugly now.

* SystemTimer.cs: A simple heartbeat timer to regulate CPU execution.

* Program.cs:
* PERQemu.csproj:
* Readme-source.txt:
* CommandExecutor.cs:
* CommandProcessor.cs: Updates and fiddling and tweaks, oh my!

* Display.cs: I'm sure if I just keep messing with it...