Add the streaming tape drive option.
Introduces support for the external QIC tape drive.  See the UserGuide.pdf
for [preliminary] information about using the tape drive. This is very much a
work-in-progress!

Bumping the version number to 0.4.8 to reflect the change.  Streamer
is still very rough but almost all of the "scaffolding" required is complete.
It should be feature complete, cleaned up and fully tested before the end
of the year...

* Readme.txt:
* ReadMe.txt:
* ChangeLog.txt:
* UserGuide.pdf:
* PERQemu.csproj:
* Display.cs:
* Streamer.txt:
* Readme-source.txt:
* FileUtilities.cs:
* Writer.cs:
* OIO.cs:
* Reader.cs:
* EventLoop.cs:
* NewIOPorts.txt:
* Events.cs:
* DeviceInfo.cs:
* qictape.png:
* PERQmedia.projitems:
* PERQSystem.cs:
* StorageDevices.db:
* Settings.cs:
* TAPFormatter.cs:
* DebugCommands.cs:
* DeviceGeometry.cs:
* OptionBoard.cs:
* AssemblyInfo.cs:
* Configurator.cs:
* DevicePerformance.cs:
* Configuration.cs:
* ConfigCommands.cs:
* StorageCommands.cs:
* SettingsCommands.cs:
* Sidewinder.cs:
* QICTapeController.cs: Add streamer tape option.

* Scheduler.cs:
* FloppyDisk.cs: A bit of lint.