Tape writes, redux
So the streaming write protocol is effing bananas.  And the PERQ's
microcode seems to do things explicitly that it doesn't need to, which
actually makes the emulation harder than it needs to be.  As in, I have
to actually implement all of it and not just the easy parts. :-)  Snapshot
for further debugging; reads to follow shortly.

* Streamer.txt: Update.

* PERQemu.csproj:
* BlockBuffer.cs: Increase wieldiness. Wieldification. Wieldinoscitude.
  Pull the block buffers out since the file is getting huge and make them
  a little more efficient (less pointless data copying) and a little smarter.
  Wish I could say the same, personally.

* QICTapeController.cs: Start cleaning up debugging stuff.

* TAPFormatter.cs: Doh. Let the patching begin!

* Sector.cs: Inevitability.

* Sidewinder.cs: You are in a maze of twisty little passages, all alike.