Cleanup the tape debugging stuff.
Tape stuff is pretty solid.  Still a ton of testing to do, but I think it's
stable enough to do some housekeeping, tag it, and push out v0.4.8
as a new package.  Stut and some tape images to follow, Real Soon Now.

* CartridgeTape.cs: Turn the debugging noise way down

* Sidewinder.cs: Turn the debugging noise way down, and reorganize
the Write loop to be a little less chaotic. Like, 4% less.

* UserGuide.pdf: Small updates for the tape stuff.